requirements.txt
docopt==0.6.2
nose2==0.6.5
py==1.4.31
pytest==3.0.3
PyYAML==3.12
six==1.10.0
